% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betareg.R
\name{betareg_model}
\alias{betareg_model}
\alias{betareg}
\alias{betareg_likelihood}
\alias{betareg_gradient}
\alias{sum_weighted_betareg_lik}
\alias{sum_weighted_betareg_grad}
\title{Compute the Beta regression model}
\usage{
betareg_likelihood(w, H, data, lambda = 1/2, is_NLL = FALSE)

betareg_gradient(w, H, data, lambda = 1/2, is_NLL = FALSE)

sum_weighted_betareg_lik(w, x, des_mat, post_prob, lambda = 1/2,
  is_NLL = TRUE)

sum_weighted_betareg_grad(w, x, des_mat, post_prob, lambda = 1/2,
  is_NLL = TRUE)
}
\arguments{
\item{w}{A vector of parameters (i.e. coefficients of the basis functions)}

\item{H}{The \code{L x M} matrix design matrix, where L is the number of
observations and M the number of basis functions.}

\item{data}{An \code{L x 2} matrix containing in the 1st column are the
observations, in the 2nd column are the proportions. Each row corresponds
to each row of the design matrix.}

\item{lambda}{The complexity penalty coefficient for penalized regression.}

\item{is_NLL}{Logical, indicating if the Negative Log Likelihood should be
returned.}

\item{x}{A list of elements of length N, where each element is an L x 2
matrix of observations, where 1st column contains the locations. The 2nd
column contains the proportions.}

\item{des_mat}{A list of length N, where each element contains the \code{L x
M} design matrices, where L is the number of observations and M the number
of basis functions.}

\item{post_prob}{A vector of length N containing the posterior probabilities
for each element of list x, respectively.}
}
\value{
Either the Beta regression log likelihood or the gradient.
}
\description{
These functions evaluate the Beta regression model likelihood
  and gradient.There are also functions to compute the sum of Beta regression
  likelihoods and weighted sum of BPR likelihoods. They are written in C++
  for efficiency (not yet!!).
}
\section{Mathematical formula}{
 The Beta distributed Probit Regression log
  likelihood function is computed by the following formula: \deqn{log p(y |
  x, w) = \sum_{l=1}^{L} log Beta(y_{l} | t_{l}, \Phi(w^{T}h(x_{l})))} where
  h(x_{l}) are the basis functions, and Beta is reparametrized to contain
  mean and dispersion parameters.
}

\seealso{
\code{\link{eval_functions}}, \code{\link{betareg_optimize}}
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
