% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_diff_test_wrap.R
\name{bpr_diff_test_wrap}
\alias{bpr_diff_test_wrap}
\title{Perform likelihood ratio test between methylation samples}
\usage{
bpr_diff_test_wrap(x, w = NULL, basis = NULL, opt_method = "CG",
  opt_itnmax = 100, is_parallel = TRUE, no_cores = NULL)
}
\arguments{
\item{x}{The binomial distributed observations. A list containing two lists
for control and treatment samples. Each list has elements of length N,
where each element is an L x 3 matrix of observations, where 1st column
contains the locations. The 2nd and 3rd columns contain the total reads and
number of successes at the corresponding locations, repsectively. See
\code{\link{process_haib_caltech_wrap}} on a possible way to get this data
structure.}

\item{w}{Optional vector of initial parameter / coefficient values.}

\item{basis}{Optional basis function object, default is an 'rbf' object, see
\code{\link{create_rbf_object}}.}

\item{opt_method}{The optimization method to be used. See
\code{\link[stats]{optim}} for possible methods. Default is "CG".}

\item{opt_itnmax}{Optional argument giving the maximum number of iterations
for the corresponding method. See \code{\link[stats]{optim}} for details.}

\item{is_parallel}{Logical, indicating if code should be run in parallel.}

\item{no_cores}{Number of cores to be used, default is max_no_cores - 2.}
}
\value{
A 'bpr_test' object which, in addition to the input
  parameters, consists of the following variables: \itemize{ \item{
  \code{W_opt}: An Nx(2M+2) matrix with the optimized parameter values. Each
  row of the matrix corresponds to the concatenated coefficients of the
  methylation profiles from both samples. The columns are of the same length
  as the concatenated parameter vector [w_contr, w_treat] (i.e. number of
  basis functions). } \item{ \code{Mus}: A list containing two matrices of
  size N x M with the RBF centers for each sample, if basis object is
  \code{\link{create_rbf_object}}, otherwise NULL.} \item{train}: The
  training data. \item{test}: The test data. \item \code{gex_model}: The
  fitted regression model. \item \code{train_pred} The predicted values for
  the training data. \item \code{test_pred} The predicted values for the test
  data. \item \code{train_errors}: The training error metrics. \item
  \code{test_errors}: The test error metrics.}
}
\description{
\code{bpr_diff_test_wrap} is a function that wraps all the necessary
subroutines for performing statistical testing between methylation samples
using the likelihood ratio test.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{bpr_optimize}}, \code{\link{create_basis}},
  \code{\link{eval_functions}}, \code{\link{train_model_gex}},
  \code{\link{predict_model_gex}}
}

