% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpr_cluster_wrap.R
\name{bpr_EM_fast}
\alias{bpr_EM_fast}
\title{Cluster methylation profiles fast method}
\usage{
bpr_EM_fast(x, H, K = 2, pi_k = rep(1/K, K), w = NULL,
  em_max_iter = 100, epsilon_conv = 1e-05, lambda = 1/2,
  opt_method = "CG", opt_itnmax = 100, is_verbose = FALSE)
}
\arguments{
\item{x}{Observations, stored in a list object.}

\item{H}{The design matrix H}

\item{K}{Integer denoting the number of clusters K.}

\item{pi_k}{Vector of length K, denoting the mixing proportions.}

\item{w}{A MxK matrix, where each column consists of the basis function
coefficients for each corresponding cluster.}

\item{em_max_iter}{Integer denoting the maximum number of EM iterations.}

\item{epsilon_conv}{Numeric denoting the convergence parameter for EM.}

\item{lambda}{The complexity penalty coefficient for ridge regression.}

\item{opt_method}{The optimization method to be used. See
\code{\link[stats]{optim}} for possible methods. Default is "CG".}

\item{opt_itnmax}{Optional argument giving the maximum number of iterations
for the corresponding method. See \code{\link[stats]{optim}} for details.}

\item{is_verbose}{Logical, print results during EM iterations.}
}
\description{
Efficient implementation for clustering methylation profiles
  using the EM algorithm, given a design matrix H.
}
