% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lm_data}
\alias{lm_data}
\title{Synthetic data for BPRMeth package}
\format{A list with 600 elements, where each element element is an L x 2
  matrix of observations, where: \describe{ \item{1st column}{locations of
  obseravtions} \item{2nd column}{outputs at corresponding locations} }}
\usage{
lm_data
}
\value{
Synthetic array-based methylation data
}
\description{
A synthetic dataset containinig 600 entries for basis function linear
regression.
}
\keyword{datasets}
