% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_bpr_cluster_wrap.R
\name{optim_regions}
\alias{optim_regions}
\title{Optimize a promoter regions across cells, which are weighted by the
responsibilities of belonging to each cluster.}
\usage{
optim_regions(x, H, w, K, opt_method = opt_method, opt_itnmax, post_prob,
  lambda)
}
\description{
Optimize a promoter regions across cells, which are weighted by the
responsibilities of belonging to each cluster.
}
