% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_lm.R
\name{predict.basis_lm}
\alias{predict.basis_lm}
\title{Make predictions using the Basis Linear Model}
\usage{
\method{predict}{basis_lm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{\link{basis_lm}}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{...}{Optional additional parameters.}
}
\value{
\code{predict.blm} produces a vector of predictions.
}
\description{
S3 method for class basis_lm, that makes predictions for \code{newdata} using
the Basis Linear Model. This funciton is similar to \code{\link[stats]{lm}},
however currently it does not provide the \code{\link[stats]{formula}}
functionality.
}
