% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr.R
\name{predict.blr}
\alias{predict.blr}
\title{Make predictions using the Basis Linear Model}
\usage{
\method{predict}{blr}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{\link{blr}}.}

\item{newdata}{An optional data frame in which to look for variables with
which to predict. If omitted, the fitted values are used.}

\item{...}{Optional additional parameters.}
}
\value{
\code{predict.blr} produces a vector of predictions.
}
\description{
S3 method for class blr, that makes predictions for \code{newdata} using
the Basis Linear Model. This funciton is similar to \code{\link[stats]{lm}},
however currently it does not provide the \code{\link[stats]{formula}}
functionality.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
