% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification_model_diff_gex.R
\name{predict_classification_model}
\alias{predict_classification_model}
\title{Predict gene expression model from methylation profiles}
\usage{
predict_classification_model(model, test, is_summary = TRUE)
}
\arguments{
\item{model}{The fitted regression model, i.e. the output of
\code{\link{train_model_gex}}.}

\item{test}{The testing data.}

\item{is_summary}{Logical, print the summary statistics.}
}
\value{
A list containing the following elements: \itemize{ \item
  \code{test_pred}: The predicted values for the test data. \item
  \code{test_errors}: The test error metrics. }
}
\description{
\code{predict_classification_model} makes predictions of gene expression
levels using a model trained on higher order methylation features extracted
from specific genomic regions.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
\seealso{
\code{\link{train_model_gex}}
}

