% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_lm.R
\name{print.basis_lm}
\alias{print.basis_lm}
\title{Print the output of the Basis Linear Model}
\usage{
\method{print}{basis_lm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{basis_lm}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class basis_lm, that prints the output of the Basis Linear Model
in a similar way to \code{\link[stats]{lm}}.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
