% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr.R
\name{print.blr}
\alias{print.blr}
\title{Print the output of the Basis Linear Model}
\usage{
\method{print}{blr}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{blr}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class basis_lm, that prints the output of the Basis Linear Model
in a similar way to \code{\link[stats]{lm}}.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
