% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_lm.R
\name{print.summary.basis_lm}
\alias{print.summary.basis_lm}
\title{Print summary output of the Basis Linear Model}
\usage{
\method{print}{summary.basis_lm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{summary.basis_lm}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class summary.basis_lm, that prints the summary of the Basis
Linear Model in a similar way to \code{\link[stats]{lm}}.
}
