% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr.R
\name{print.summary.blr}
\alias{print.summary.blr}
\title{Print summary output of the Basis Linear Model}
\usage{
\method{print}{summary.blr}(x, ...)
}
\arguments{
\item{x}{Object of class \code{\link{summary.blr}}.}

\item{...}{Optional additional parameters.}
}
\description{
S3 method for class summary.blr, that prints the summary of the Basis
Linear Model in a similar way to \code{\link[stats]{lm}}.
}
