% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_lm.R
\name{summary.basis_lm}
\alias{summary.basis_lm}
\title{Summary output of the Basis Linear Model}
\usage{
\method{summary}{basis_lm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{\link{basis_lm}}.}

\item{...}{Optional additional parameters.}
}
\value{
A summary.basis_lm object
}
\description{
S3 method for class basis_lm, that computes the summary of the Basis Linear
Model in a similar way to \code{\link[stats]{lm}}.
}
