% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blr.R
\name{summary.blr}
\alias{summary.blr}
\title{Summary output of the Basis Linear Model}
\usage{
\method{summary}{blr}(object, ...)
}
\arguments{
\item{object}{Object of class \code{\link{blr}}.}

\item{...}{Optional additional parameters.}
}
\value{
A summary.blr object
}
\description{
S3 method for class blr, that computes the summary of the Basis Linear
Model in a similar way to \code{\link[stats]{lm}}.
}
\author{
C.A.Kapourani \email{C.A.Kapourani@ed.ac.uk}
}
